refactor: move shared common/agent types out of config crate and add agent-specific config blocks (#9)
* refactor: move common and most agent types into rust/nomad-types

* fix: add metrics port to AgentConfig

* fix: add metrics ports to json configs

* feature: add agent-specific configs

* fix: make agent config fields pub

* fix: index_only required field

* fix: make metric port optional

* chore: make structs partial eq for agent config checks

* fix: make uints consistent with rust repo

* fix: use db_path since cannot create items in usr/share

* nits: affix instead of paste, db_path --> db

* fix: use deser_nomad_<x> and update changelog

+++ src/contracts.rs
@@ -2,17 +2,16 @@
 
 use std::collections::HashMap;
 
-use crate::common::{NomadIdentifier, Proxy};
-
-use crate::common::deser_nomad_number;
+use nomad_types::deser_nomad_u32;
+use nomad_types::{NomadIdentifier, Proxy};
 
 /// Evm Core Contracts
 #[derive(Default, Debug, Clone, serde::Serialize, serde::Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct EvmCoreContracts {
     /// Contract Deploy Height
-    #[serde(default, deserialize_with = "deser_nomad_number")]
-    pub deploy_height: u64,
+    #[serde(default, deserialize_with = "deser_nomad_u32")]
+    pub deploy_height: u32,
     /// UBC address
     pub upgrade_beacon_controller: NomadIdentifier,
     /// XApp Connection Manager address
